Table values are essentially ints, not floats

+++ coronavirus/coronavirus.py
@@ -253,7 +253,7 @@ def compose_dataframe_summary(cases, deaths):
     df["daily new deaths"] = deaths.diff()
 
     # drop first row with nan -> otherwise ints are shows as float in table
-    df = df.dropna()
+    df = df.dropna().astype(int)
 
     # change index: latest numbers shown first
     df = df[::-1]
